requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── LLM & Embeddings ────────────────────────────────────────────────────
groq==0.15.0
sentence-transformers>=3.3.0
transformers>=4.47.0
torch>=2.0.0

# ── Vector Store ─────────────────────────────────────────────────────────
faiss-cpu==1.9.0.post1
numpy>=1.26.0

# ── Resilience ───────────────────────────────────────────────────────────
tenacity==9.0.0

# ── HTTP ─────────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Testing ──────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0

# ── Dev / Linting (optional) ────────────────────────────────────────────
python-dotenv==1.0.1
